+++ src/components/PostCards/PostCard.js
@@ -1,27 +1,32 @@
 import React, { Component } from 'react'
 /* firebase and firestore */
 import { addDoc, collection, Timestamp } from 'firebase/firestore'
-import { db } from '../firebase'
+import { db } from '../../firebase'
 import { Card, Button } from 'react-bootstrap'
-import SharingButtons from './Sharing/SharingButtons'
+import SharingButtons from '../Sharing/SharingButtons'
+import { CopyToClipboard } from 'react-copy-to-clipboard'
 import moment from 'moment'
 
+/* toast */
 import { ToastContainer, toast } from 'react-toastify'
-export class PostCardMovies extends Component {
+
+export class PostCard extends Component {
   constructor(props) {
     super(props)
     this.state = {
       author: this.props.post.byline,
-      date: this.props.post.publication_date,
+      date: this.props.post.published_date,
       createdAt: Timestamp.now(),
-      description: this.props.post.summary_short,
-      section: 'movies',
-      title: this.props.post.display_title,
-      url: this.props.post.link.url,
+      description: this.props.post.abstract,
+      section: this.props.post.section,
+      title: this.props.post.title,
+      url: this.props.post.url,
       user: this.props.user.multiFactor.user.uid,
       source: 'New York Times',
+      comments: '',
       favorites: this.props.favorites,
       active: true,
+      copied: false,
     }
   }
 
@@ -33,33 +38,32 @@ export class PostCardMovies extends Component {
       /* 'New York Times' */
       this.props.user.multiFactor.user.uid
     )
-    const saveFavorite = async (props) => {
+    const saveFavorite = async () => {
       try {
         await addDoc(favoritesCollectionRef, {
           author: this.props.post.byline,
-          date: this.props.post.publication_date,
+          date: this.props.post.published_date,
           createdAt: Timestamp.now(),
-          description: this.props.post.summary_short,
-          section: 'movies',
-          title: this.props.post.display_title,
-          url: this.props.post.link.url,
+          description: this.props.post.abstract,
+          section: this.props.post.section,
+          title: this.props.post.title,
+          url: this.props.post.url,
           user: this.props.user.multiFactor.user.uid,
           source: 'New York Times',
+          comments: '',
         })
-        /* console.log('favorite added')
-        console.log(this.props.post.byline) */
+
         toast('favorite added')
         this.setState({
           active: false,
         })
       } catch (err) {
-        /* console.log(err) */
         toast(err)
       }
     }
 
     let favoritedItem = this.props.favorites.filter(
-      (favorite) => favorite.title === this.props.post.display_title
+      (favorite) => favorite.title == this.props.post.title
     )
 
     return (
@@ -70,17 +74,24 @@ export class PostCardMovies extends Component {
         >
           <Card.Body>
             {' '}
-            <div className='title-card'>{this.props.post.display_title}</div>
+            <div className='title-card'>{this.props.post.title}</div>
             {this.props.post.multimedia ? (
               <Card.Img
-                className='img-movies'
-                src={this.props.post.multimedia.src}
-                alt={this.props.post.headline}
+                className='img'
+                src={this.props.post.multimedia[0].url}
+                alt={this.props.post.caption}
               />
             ) : null}
-            <div className='subtitle'>{this.props.post.summary_short}</div>
+            {this.props.post.book_image ? (
+              <Card.Img
+                className='img book-img'
+                src={this.props.post.book_image}
+                alt={this.props.post.description}
+              />
+            ) : null}
+            <div className='subtitle'>{this.props.post.abstract}</div>
             <Card.Text className='author-date'>
-              <span>By {this.props.post.byline}.</span>
+              <span>{this.props.post.byline}</span>{' '}
               <span>
                 Published:{' '}
                 {moment(this.props.post.published_date).format('MMMM DD, YYYY')}
@@ -88,14 +99,15 @@ export class PostCardMovies extends Component {
             </Card.Text>
             <div className='d-flex align-items-center justify-content-center'>
               <Button
-                className='btn read-more btn-outline-light mx-2'
-                variant='link'
+                className='btn read-more'
+                variant='btn btn-outline-light mx-2'
               >
-                <a href={this.props.post.link.url} className='myLink'>
+                <a href={this.props.post.url} className='myLink'>
                   Read more
                 </a>
                 .
               </Button>
+
               <Button
                 className='btn read-more'
                 variant='btn btn-outline-light mx-2'
@@ -109,9 +121,20 @@ export class PostCardMovies extends Component {
             </div>
           </Card.Body>
 
-          <Card.Footer>
-            <SharingButtons url={this.props.post.link.url} />
+          <Card.Footer className='d-flex align-items-center justify-content-center'>
+            <SharingButtons url={this.props.post.url} />
+            <CopyToClipboard
+              text={this.props.post.url}
+              onCopy={() => this.setState({ copied: true })}
+            >
+              <img
+                src='../assets/icons/clipboard.svg'
+                alt='read more'
+                className='sharing-icon'
+              />
+            </CopyToClipboard>
           </Card.Footer>
+          <div></div>
           <ToastContainer
             position='bottom-right'
             type='info'
@@ -131,4 +154,4 @@ export class PostCardMovies extends Component {
   }
 }
 
-export default PostCardMovies
+export default PostCard